+++ triplea/service/repository/pipeline_flag.py
@@ -422,7 +422,11 @@ def go_affiliation_mining(method: str = "Simple",
                         updated_article
                     )
                     persist.update_article_by_id(updated_article, id)
-
+                elif method == "TitipataIntegrated":
+                    updated_article = state_manager.affiliation_mining_titipata_integration(
+                        updated_article
+                    )
+                    persist.update_article_by_id(updated_article, id)
             elif current_state == 1:
                 pass